requirements.txt
altair
numpy
pandas
pydeck
streamlit
haystack-ai
datasets>=2.6.1
sentence-transformers>=2.2.0
streamlit
transformers